Introduce junit5 integration test extension #3828
- junit4 rule logic moved to IntegrationTestSupport class
- junit4 rule uses now the integration test support
- junit5 extension IntegrationTestExtension created
  + WithScenario annotation
- junit5 extension uses also IntegrationTestSupport class
