+++ sechub-integrationtest/src/main/java/com/mercedesbenz/sechub/integrationtest/api/WithTestScenario.java
@@ -1,0 +1,20 @@
+package com.mercedesbenz.sechub.integrationtest.api;
+
+import java.lang.annotation.Documented;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Inherited;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+import com.mercedesbenz.sechub.integrationtest.internal.TestScenario;
+
+@Target({ ElementType.TYPE })
+@Retention(RetentionPolicy.RUNTIME)
+@Documented
+@Inherited
+public @interface WithTestScenario {
+
+    Class<? extends TestScenario> value();
+
+}
